Add first draft code
Add a summary json file that lists all of the books. I needed this file
to look through the books. I added the basic code for grids and the
summary table. Inspiration for the grid from
https://medium.com/@andybarefoot/d3-and-css-grid-with-expanding-content-3c8aaf783cb1

I think this could be a cool way to summarize all of the books and then
have the table on a separate tab. I agree with our reviewers from class
that the table is too big for a summary look. In the grid, the tiles are
currently sized by total reviews. We could allow the user to set different
tile sizes.