Fixing the Book Grid. It now appears correctly.
Probably need to discuss, but I have spent a week on ways to get the
scales to work. Since the range of the data columns is so large, the
bars look really small and not visible for most books and only appears
large for a few books. I think we should move to percentages.